+++ src/main/java/net/alvin/infinityforge/client/render/ModItemRenderers.java
@@ -18,22 +18,24 @@ public class ModItemRenderers {
 
     public static void register() {
         for (InfinityStoneType stoneType : InfinityStoneTypeRegistry.STONE_TYPE_REGISTRY) {
-            Identifier typeId = InfinityStoneTypeRegistry.STONE_TYPE_REGISTRY.getId(stoneType);
+            String stoneName = InfinityStoneTypeRegistry.getNameFromType(stoneType, " Stone", true);
+            Identifier stoneId = InfinityStoneTypeRegistry.getIdentifierFromType(stoneType, "_stone");
+            Item stoneItem = Registries.ITEM.get(stoneId);
 
-            Identifier stoneId = typeId.withPath(typeId.getPath() + "_stone");
-            Item stone = Registries.ITEM.get(stoneId);
-            InfinityForge.LOGGER.info("Registering renderer for: " + stoneId + " item: " + stone);
-            BuiltinItemRendererRegistry.INSTANCE.register(stone,
+            InfinityForge.LOGGER.info("Registering renderer for Item: {} ({})", stoneName, stoneId.toString());
+            BuiltinItemRendererRegistry.INSTANCE.register(stoneItem,
                     (stack, mode,
                      matrices, vertexConsumers,
                      light, overlay) -> STONE_RENDERER.render(stack, mode,
                             matrices, vertexConsumers,
                             light, overlay, stoneType));
 
-            Identifier tesseractId = typeId.withPath(typeId.getPath() + "_tesseract");
-            Item tesseract = Registries.ITEM.get(tesseractId);
-            InfinityForge.LOGGER.info("Registering renderer for: " + tesseractId + " item: " + tesseract);
-            BuiltinItemRendererRegistry.INSTANCE.register(tesseract,
+            String tesseractName = InfinityStoneTypeRegistry.getNameFromType(stoneType, " Tesseract", true);
+            Identifier tesseractId = InfinityStoneTypeRegistry.getIdentifierFromType(stoneType, "_tesseract");
+            Item tesseractItem = Registries.ITEM.get(tesseractId);
+
+            InfinityForge.LOGGER.info("Registering renderer for Item: {} ({})", tesseractName, tesseractId.toString());
+            BuiltinItemRendererRegistry.INSTANCE.register(tesseractItem,
                     (stack, mode,
                      matrices, vertexConsumers,
                      light, overlay) -> TESSERACT_RENDERER.render(stack, mode,
@@ -43,7 +45,6 @@ public class ModItemRenderers {
 
         ModelLoadingPlugin.register(pluginContext -> pluginContext.addModels(InfinityGauntletRenderer.GAUNTLET_MODEL_2D));
         BuiltinItemRendererRegistry.INSTANCE.register(ModItems.INFINITY_GAUNTLET, GAUNTLET_RENDERER::render);
-
         BuiltinItemRendererRegistry.INSTANCE.register(ModItems.FAKE_ITEM, FAKE_ITEM_RENDERER::render);
     }
 }